# docs/openapi.yaml
openapi: 3.1.0
info:
  title: LMS V1 API
  version: 1.0.0
  description: Canonical single-owner LMS contract. NestJS is the authorization authority.
  license:
    name: UNLICENSED
    identifier: LicenseRef-Proprietary
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
...
security: []
paths:
...
  /me:
    get:
      tags: [Account]
      operationId: getMe
      summary: Get the current subject profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [Account]
      operationId: updateMe
      summary: Update the current subject profile
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              minProperties: 1
              properties:
                firstName: { type: string, minLength: 1, maxLength: 100 }
                lastName: { type: string, minLength: 1, maxLength: 100 }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "401": { $ref: "#/components/responses/Unauthorized" }
...
  /me/courses:
    get:
      tags: [Learning]
      operationId: listMyCourses
      summary: List courses available to the current student
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Courses covered by subject enrollments
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CoursePage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
  /me/courses/{courseId}:
    get:
      tags: [Learning]
      operationId: getMyCourse
      summary: Get an authorized course curriculum
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/CourseId"
      responses:
        "200":
          description: Authorized curriculum
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CourseWithCurriculum" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /me/lessons/{lessonId}:
    get:
      tags: [Learning]
      operationId: getMyLesson
      summary: Get an authorized lesson
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/LessonId"
      responses:
        "200":
          description: Authorized lesson without storage credentials or keys
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Lesson" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /me/lessons/{lessonId}/progress:
    put:
      tags: [Learning]
      operationId: updateMyLessonProgress
      summary: Save current-subject lesson progress
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/LessonId"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProgressRequest" }
      responses:
        "200":
          description: Saved progress
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LessonProgress" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /owner/lessons/{lessonId}/video-uploads:
    post:
      tags: [Owner videos]
      operationId: initiateVideoUpload
      summary: Authorize a direct private video upload
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/LessonId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/VideoUploadRequest" }
      responses:
        "201":
          description: Direct private upload authorized
          content:
            application/json:
              schema: { $ref: "#/components/schemas/VideoUpload" }
        "400": { $ref: "#/components/responses/ValidationFailed" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Unprocessable" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /owner/lessons/{lessonId}/resources:
    post:
      tags: [Owner courses]
      operationId: initiateLessonResourceUpload
      summary: Authorize a direct private lesson-resource upload
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/LessonId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LessonResourceUploadRequest" }
      responses:
        "201":
          description: Temporary private upload authorization
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LessonResourceUpload" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /owner/lesson-resources/{resourceId}/complete:
    post:
      tags: [Owner courses]
      operationId: completeLessonResourceUpload
      summary: Verify a private lesson-resource upload
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Resource ready
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LessonResource" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /me/lesson-resources/{resourceId}/download:
    post:
      tags: [Learning]
      operationId: createLessonResourceDownload
      summary: Authorize a short-lived download for an enrolled subject
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "201":
          description: Five-minute private download authorization
          content:
            application/json:
              schema: { $ref: "#/components/schemas/LessonResourceDownload" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /owner/video-uploads/{uploadId}/complete:
    post:
      tags: [Owner videos]
      operationId: completeVideoUpload
      summary: Verify upload completion and queue processing
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/UploadId"
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "202":
          description: Upload verified and processing queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Video" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /owner/video-uploads/{uploadId}/parts/{partNumber}:
    post:
      tags: [Owner videos]
      operationId: createVideoUploadPartUrl
      summary: Authorize one direct multipart upload part
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/UploadId"
        - name: partNumber
          in: path
          required: true
          schema: { type: integer, minimum: 1, maximum: 10000 }
      responses:
        "201":
          description: Temporary part upload URL
          content:
            application/json:
              schema:
                type: object
                required: [uploadUrl]
                properties: { uploadUrl: { type: string, format: uri } }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /owner/video-uploads/{uploadId}/multipart-complete:
    post:
      tags: [Owner videos]
      operationId: completeMultipartVideoUpload
      summary: Complete, verify, and queue a multipart video upload
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/UploadId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompleteMultipartUploadRequest" }
      responses:
        "202":
          description: Multipart upload verified and processing queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Video" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/Unprocessable" }
  /owner/videos/{videoId}:
    get:
      tags: [Owner videos]
      operationId: getVideo
      summary: Get video processing status
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/VideoId"
      responses:
        "200":
          description: Video processing status
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Video" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: lms_refresh
  parameters:
    Cursor:
      name: cursor
      in: query
      schema: { type: string, maxLength: 512 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, minLength: 16, maxLength: 128 }
    CourseId: { name: courseId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
...
    LessonId: { name: lessonId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
...
    VideoId: { name: videoId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
    UploadId: { name: uploadId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
    ResourceId: { name: resourceId, in: path, required: true, schema: { $ref: "#/components/schemas/Uuid" } }
  responses:
    ValidationFailed:
      description: Request validation failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Authentication failed
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Authorization or account policy denied access
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Resource not found or concealed by authorization
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Conflict:
      description: State or uniqueness conflict
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unprocessable:
      description: Semantically invalid upload or processing input
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    RateLimited:
      description: Rate limit exceeded
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
    Uuid: { type: string, format: uuid }
    Timestamp: { type: string, format: date-time }
    Error:
      type: object
      additionalProperties: false
      required: [statusCode, code, message, requestId]
      properties:
        statusCode: { type: integer, minimum: 400, maximum: 599 }
        code: { type: string, pattern: "^[A-Z][A-Z0-9_]+$" }
        message: { type: string }
        details: { type: object, additionalProperties: true }
        requestId: { $ref: "#/components/schemas/Uuid" }
...
    User:
      type: object
      additionalProperties: false
      required: [id, email, firstName, lastName, status, roles, createdAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        email: { type: string, format: email }
        firstName: { type: string }
        lastName: { type: string }
        status: { type: string, enum: [PENDING_VERIFICATION, ACTIVE, SUSPENDED, DISABLED] }
        roles: { type: array, items: { type: string, enum: [OWNER, STUDENT] } }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
...
    Course:
      type: object
      additionalProperties: false
      required: [id, title, slug, description, status, createdAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        title: { type: string }
        slug: { type: string }
        description: { type: string }
        thumbnailUrl: { type: [string, "null"], format: uri }
        status: { type: string, enum: [DRAFT, PUBLISHED, ARCHIVED] }
        publishedAt: { type: [string, "null"], format: date-time }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }
    CoursePage:
      type: object
      required: [items, nextCursor]
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Course" } }
        nextCursor: { type: [string, "null"] }
    CourseWithCurriculum:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          required: [sections]
          properties:
            sections: { type: array, items: { $ref: "#/components/schemas/Section" } }
...
    Section:
      type: object
      required: [id, courseId, title, sortOrder, lessons]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        courseId: { $ref: "#/components/schemas/Uuid" }
        title: { type: string }
        description: { type: [string, "null"] }
        sortOrder: { type: integer, minimum: 0 }
        lessons: { type: array, items: { $ref: "#/components/schemas/Lesson" } }
...
    Lesson:
      type: object
      required: [id, sectionId, title, type, sortOrder]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        sectionId: { $ref: "#/components/schemas/Uuid" }
        title: { type: string }
        description: { type: [string, "null"] }
        type: { type: string, enum: [VIDEO, TEXT] }
        textContent: { type: [string, "null"] }
        sortOrder: { type: integer, minimum: 0 }
        progress: { oneOf: [{ $ref: "#/components/schemas/LessonProgress" }, { type: "null" }] }
        resources:
          type: array
          items: { $ref: "#/components/schemas/LessonResource" }
    LessonResource:
      type: object
      additionalProperties: false
      required: [id, lessonId, title, filename, mimeType, sizeBytes, createdAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        lessonId: { $ref: "#/components/schemas/Uuid" }
        title: { type: string }
        filename: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: string, pattern: "^[1-9][0-9]*$" }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
    LessonResourceUploadRequest:
      type: object
      additionalProperties: false
      required: [title, filename, mimeType, sizeBytes]
      properties:
        title: { type: string, minLength: 1, maxLength: 200 }
        filename: { type: string, minLength: 1, maxLength: 255 }
        mimeType:
          type: string
          enum: [application/pdf, text/plain, application/zip, application/vnd.openxmlformats-officedocument.wordprocessingml.document]
        sizeBytes: { type: integer, minimum: 1, maximum: 104857600 }
    LessonResourceUpload:
      allOf:
        - $ref: "#/components/schemas/LessonResource"
        - type: object
          required: [uploadUrl, expiresAt]
          properties:
            uploadUrl: { type: string, format: uri }
            expiresAt: { $ref: "#/components/schemas/Timestamp" }
    LessonResourceDownload:
      type: object
      additionalProperties: false
      required: [id, title, filename, mimeType, sizeBytes, downloadUrl, expiresIn]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        title: { type: string }
        filename: { type: string }
        mimeType: { type: string }
        sizeBytes: { type: string, pattern: "^[1-9][0-9]*$" }
        downloadUrl: { type: string, format: uri }
        expiresIn: { type: integer, enum: [300] }
...
    ProgressRequest:
      type: object
      additionalProperties: false
      required: [positionSeconds, watchedSeconds]
      properties:
        positionSeconds: { type: integer, minimum: 0 }
        watchedSeconds: { type: integer, minimum: 0 }
    LessonProgress:
      type: object
      required: [lessonId, lastPositionSeconds, maximumPositionSeconds, watchedSeconds, percentage]
      properties:
        lessonId: { $ref: "#/components/schemas/Uuid" }
        lastPositionSeconds: { type: integer, minimum: 0 }
        maximumPositionSeconds: { type: integer, minimum: 0 }
        watchedSeconds: { type: integer, minimum: 0 }
        percentage: { type: number, minimum: 0, maximum: 100 }
        completedAt: { type: [string, "null"], format: date-time }
...
    VideoUploadRequest:
      type: object
      additionalProperties: false
      required: [filename, mimeType, sizeBytes]
      properties:
        filename: { type: string, minLength: 1, maxLength: 255 }
        mimeType: { type: string, enum: [video/mp4, video/quicktime, video/x-matroska] }
        sizeBytes: { type: integer, minimum: 1 }
    VideoUpload:
      type: object
      additionalProperties: false
      required: [id, videoId, status, uploadMode, uploadUrl, partSizeBytes, expiresAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        videoId: { $ref: "#/components/schemas/Uuid" }
        status: { type: string, enum: [PENDING, UPLOADING] }
        uploadMode: { type: string, enum: [SINGLE, MULTIPART] }
        uploadUrl: { type: [string, "null"], format: uri }
        partSizeBytes: { type: [integer, "null"], minimum: 5242880 }
        expiresAt: { $ref: "#/components/schemas/Timestamp" }
    CompleteMultipartUploadRequest:
      type: object
      additionalProperties: false
      required: [parts]
      properties:
        parts:
          type: array
          minItems: 1
          maxItems: 10000
          items:
            type: object
            additionalProperties: false
            required: [partNumber, etag]
            properties:
              partNumber: { type: integer, minimum: 1, maximum: 10000 }
              etag: { type: string, minLength: 1, maxLength: 130 }
    Video:
      type: object
      additionalProperties: false
      required: [id, lessonId, status, sourceFilename, sourceSizeBytes, isCurrent, createdAt, updatedAt]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        lessonId: { $ref: "#/components/schemas/Uuid" }
        status: { type: string, enum: [UPLOADING, UPLOADED, QUEUED, PROCESSING, READY, FAILED, DELETING, DELETED] }
        sourceFilename: { type: string }
        sourceSizeBytes: { type: integer, minimum: 1 }
        durationSeconds: { type: [integer, "null"], minimum: 0 }
        width: { type: [integer, "null"], minimum: 1 }
        height: { type: [integer, "null"], minimum: 1 }
        isCurrent: { type: boolean }
        processingError: { type: [string, "null"] }
        createdAt: { $ref: "#/components/schemas/Timestamp" }
        updatedAt: { $ref: "#/components/schemas/Timestamp" }